some ui changes

+++ chat-bot-ui/src/context-api/ChatContext.jsx
@@ -1,5 +1,6 @@
 import axios from "axios";
-import { createContext, useEffect, useState } from "react";
+import { createContext, useContext, useEffect, useState } from "react";
+import { AuthContext } from "./AuthContext";
 
 export const ChatContext = createContext();
 
@@ -13,15 +14,17 @@ export const ChatContextProvider = ({children}) => {
     const [currentUser,setCurrentUser] = useState(null);
     const [createdChatId,setCreatedChatId] = useState(null);
 
-    const token = localStorage.getItem('token') || null;
+    const {accessToken, authReady} = useContext(AuthContext);
+
     
     const getChatList = async () => {
+        if (!accessToken) return; 
         try {
             const {data} = await axios.get(backendUrl+"/api/v1/chat",{
                 headers: {
                     "Content-Type": "application/json",
                     Accept: "application/json",
-                    Authorization: "Bearer "+token,
+                    Authorization: "Bearer "+accessToken,
                 }
             });
             setChatList(data);
@@ -31,12 +34,13 @@ export const ChatContextProvider = ({children}) => {
     }
 
     const getChatById = async (id) => {
+        if (!accessToken) return; 
         try {
             const {data} = await axios.get(backendUrl+"/api/v1/chat/"+id,{
                 headers: {
                     "Content-Type": "application/json",
                     Accept: "application/json",
-                    Authorization: "Bearer "+token,
+                    Authorization: "Bearer "+accessToken,
                 }
             });
             setChatById(data);
@@ -47,7 +51,7 @@ export const ChatContextProvider = ({children}) => {
 
     const sendMessageToAi = async (id,message) => {
         try {
-
+            if (!accessToken) return; 
             // Optimistically add your message to the chat
             const tempMessage = {
                 id: Date.now(), // temporary id
@@ -73,7 +77,7 @@ export const ChatContextProvider = ({children}) => {
                     headers: {
                         "Content-Type": "application/json",
                         Accept: "application/json",
-                        Authorization: "Bearer "+token,
+                        Authorization: "Bearer "+accessToken,
                     }
                 });
 
@@ -86,13 +90,14 @@ export const ChatContextProvider = ({children}) => {
     }
 
     const createChat = async () => {
+        if (!accessToken) return; 
         try {
             const {data} = await axios.post(backendUrl+"/api/v1/chat",{},
                 {
                 headers: {
                     "Content-Type": "application/json",
                     Accept: "application/json",
-                    Authorization: "Bearer "+token,
+                    Authorization: "Bearer "+accessToken,
                 }
             });
             console.log(data);
@@ -104,12 +109,13 @@ export const ChatContextProvider = ({children}) => {
     }
 
     const deleteChatById = async (id) => {
+        if (!accessToken) return; 
         try {
             const response = await axios.delete(backendUrl+"/api/v1/chat/"+id,{
                 headers: {
                     "Content-Type": "application/json",
                     Accept: "application/json",
-                    Authorization: "Bearer "+token,
+                    Authorization: "Bearer "+accessToken,
                 }
             });
             console.log(response);
@@ -120,6 +126,7 @@ export const ChatContextProvider = ({children}) => {
     }
 
     const updateChatTitle = async (id,newTitle) => {
+        if (!accessToken) return; 
         try {
             const response = await axios.put(backendUrl+"/api/v1/chat/"+id,
                 {
@@ -129,7 +136,7 @@ export const ChatContextProvider = ({children}) => {
                     headers: {
                         "Content-Type": "application/json",
                         Accept: "application/json",
-                        Authorization: "Bearer "+token,
+                        Authorization: "Bearer "+accessToken,
                     }
                 });
             console.log(response);
@@ -140,13 +147,14 @@ export const ChatContextProvider = ({children}) => {
     }
 
     const getCurrentUserData =  async () => {
+        if (!accessToken) return; 
         try {
             const {data} = await axios.get(backendUrl+"/api/v1/user/profile",
                 {
                     headers: {
                         "Content-Type": "application/json",
                         Accept: "application/json",
-                        Authorization: "Bearer "+token,
+                        Authorization: "Bearer "+accessToken,
                     }
                 });
             setCurrentUser(data);
@@ -155,9 +163,12 @@ export const ChatContextProvider = ({children}) => {
         }
     }
 
-    useEffect(()=>{
+    useEffect(() => {
+        if (!authReady || !accessToken) return;
 
-    },[])
+        getChatList();
+        getCurrentUserData();
+    }, [authReady, accessToken]);
 
     const value = {
         backendUrl,
